Time: 1 ms (99.85%), Space: 42.8 MB (34.98%) - LeetHub

+++ 3754-concatenate-non-zero-digits-and-multiply-by-sum-i/3754-concatenate-non-zero-digits-and-multiply-by-sum-i.java
@@ -1,10 +1,10 @@
 class Solution {
     public long sumAndMultiply(int n) {
         long x = 0;
-        long sum=0;
-        long mul = 1;
+        int sum=0;
+        int mul = 1;
         while(n>0){
-            long digit = (n%10);
+            int digit = (n%10);
             if(digit>0){
                 sum+=digit;
                 x+=digit*mul;
